added new file user_settings.py with settings page. Migrate from authorization.py. Change connect in app.py
